+++ src/main/java/ch/heigvd/amt/projet1/domain/IEntity.java
@@ -1,0 +1,4 @@
+package ch.heigvd.amt.projet1.domain;
+
+public interface IEntity {
+}
