+++ src/main/java/ch/heigvd/amt/projet1/domain/question/QuestionId.java
@@ -1,0 +1,7 @@
+package ch.heigvd.amt.projet1.domain.question;
+
+import ch.heigvd.amt.projet1.domain.Id;
+
+public class QuestionId extends Id {
+
+}
