+++ src/main/java/ch/heigvd/amt/projet1/domain/question/QuestionQuery.java
@@ -1,4 +1,4 @@
-package ch.heigvd.amt.projet1.model;
+package ch.heigvd.amt.projet1.domain.question;
 
 public class QuestionQuery { //peut être utile plus tard
 }
