--- src/main/java/ch/heigvd/amt/projet1/model/IEntity.java
@@ -1,4 +1,0 @@
-package ch.heigvd.amt.projet1.model;
-
-public interface IEntity {
-}
