--- src/main/java/ch/heigvd/amt/projet1/model/QuestionId.java
@@ -1,5 +1,0 @@
-package ch.heigvd.amt.projet1.model;
-
-public class QuestionId extends Id{
-
-}
